Change container for subnodes in SerializationInfo from std::vector to std::deque
Adding members to SerializationInfo using addMember may result to invalidation of previous references to
the object when using std::vector. This does not happen with std::deque.
Since a deque do not have a reserve method, the reserve method in SerializationInfo is removed also.
This results in a break in the api. If someone use the method, just remove it. It is not needed in a
deque.