small fix in propertiesdeserializer
When trim is enabled in propertiesdeserializer white space at the beginning
and end of a value is trimmed away. The sequence \t in a properties file
means a tab character. But this sequence should not be stripped away. This
is fixed now. Also a space, which is escaped with '\' is not stripped away
any more.