improve tcp server
When a empty ip host is passed as a listen ip to the tcp server, it listens
on all local interfaces. It used to fail if any of those interfaces fail but
now it is changed that it is ok when at least one interface is accepted as
a listener interface.