Change binary rpc server api.
Instead of loading a ssl certificate in a separate call, it is given as
a parameter to the listen call. The certificate is now specific to a listener
and not to the server any more. Since multiple listeners can be instantiated
by calling listen multiple times, it is possible to listen on ssl and non
ssl in the same server or with different certificates on different interfaces.

The backlog parameter is removed and set to 64.