improvements in binary serializer
Reduce use of iostream. streambuf is one level less.
To make it easy to use of this there are new methods in bin::Serializer.
Instead of using std::istringstream the user can use the static method
bin::Seriailzer::toString(const T&) to get a std::string, which uses now
std::streambuf, which makes it faster.