small improvement in csv serializer
The csv serializer collects the titles from objects. But the name
of the node should be used only when the category is Value.
std::optional keeps the category unset when the value of the optional is
not set. But in this case the title collector takes the name of the
object, although it may be a optional of a object. Serializing a
object collects the name of the sub nodes but not the nodes itself.